Add Easy C++ Solution from LeetCode

+++ LeetCode/C++/Easy/119. Pascals Triangle II/Solution.cpp
@@ -4,7 +4,7 @@ public:
          vector<int>anss;
          row++;
          anss.push_back(1);
-         int ans=1;
+         long long ans=1;
         for(int col=1;col<row;col++){
             ans=ans*(row-col);
             ans=ans/col;
